Reset cancellations if possible.
This also uses new subclasses of CancelledError, called
BenignCancelledError, with which we can distinguish between cancellation
detected by the optimization problem and cancellation detected by us,
the GUI.

The difference is that if we know that _we_ caught a cancellation, we
always know that it is safe to complete the cancellation.

In theory, it would be possible to not use the subclass and instead
complete the cancellation immediately as we detect it. It's not clear
which approach is better.