Add DelayedComboBox.setStableText() and fix buggy --user.
The --user parameter was unable to select users that were not associated
with the default machine. The problem was that the machine selection
combo box had a delay built in. So when the machine would be selected,
the LSA selector widget would only update several milliseconds later. By
that point, `select_user()` had already been called and failed to find
the desired user.

This is fixed by adding a method to `DelayedComboBox` that changes the
item immediately and using it instead of `setCurrentText()`.