Fix DelayedComboBox.setStableText() and setStableIndex().
The setStable*() methods used to not emit stableTextChanged directly,
but rather set the timer interval to zero. This caused the signal to
only be emitted from the event loop. This means that it was impossible
to handle this signal during initialization, which broke
`make_initial_selection()`.

We switch back to a semaphore-based solution that inhibits
`_kick_off_timer()`. This keeps in mind that no signal should be emitted
if the text doesn't actually change.